+++ tidalapi/media.py
@@ -329,10 +329,12 @@ class Track(Media):
 
         try:
             request = self.requests.request("GET", "tracks/%s" % media_id)
-        except ObjectNotFound:
-            raise ObjectNotFound("Track not found or unavailable")
-        except TooManyRequests:
-            raise TooManyRequests("Track unavailable")
+        except ObjectNotFound as e:
+            e.args = ("Track with id %s not found" % media_id,)
+            raise e
+        except TooManyRequests as e:
+            e.args = ("Track unavailable",)
+            raise e
         else:
             json_obj = request.json()
             track = self.requests.map_json(json_obj, parse=self.parse_track)
@@ -362,8 +364,9 @@ class Track(Media):
             )
         except ObjectNotFound:
             raise URLNotAvailable("URL not available for this track")
-        except TooManyRequests:
-            raise TooManyRequests("URL Unavailable")
+        except TooManyRequests as e:
+            e.args = ("URL unavailable",)
+            raise e
         else:
             json_obj = request.json()
             return cast(str, json_obj["urls"][0])
@@ -378,8 +381,9 @@ class Track(Media):
             request = self.requests.request("GET", "tracks/%s/lyrics" % self.id)
         except ObjectNotFound:
             raise MetadataNotAvailable("No lyrics exists for this track")
-        except TooManyRequests:
-            raise TooManyRequests("Lyrics unavailable")
+        except TooManyRequests as e:
+            e.args = ("Lyrics unavailable",)
+            raise e
         else:
             json_obj = request.json()
             lyrics = self.requests.map_json(json_obj, parse=Lyrics().parse)
@@ -401,8 +405,9 @@ class Track(Media):
             )
         except ObjectNotFound:
             raise MetadataNotAvailable("Track radio not available for this track")
-        except TooManyRequests:
-            raise TooManyRequests("Track radio unavailable")
+        except TooManyRequests as e:
+            e.args = ("Track radio unavailable",)
+            raise e
         else:
             json_obj = request.json()
             tracks = self.requests.map_json(json_obj, parse=self.session.parse_track)
@@ -420,8 +425,9 @@ class Track(Media):
             request = self.requests.request("GET", "tracks/%s/mix" % self.id)
         except ObjectNotFound:
             raise MetadataNotAvailable("Track radio not available for this track")
-        except TooManyRequests:
-            raise TooManyRequests("Track radio unavailable")
+        except TooManyRequests as e:
+            e.args = ("Track radio unavailable",)
+            raise e
         else:
             json_obj = request.json()
             return self.session.mix(json_obj.get("id"))
@@ -445,8 +451,9 @@ class Track(Media):
             )
         except ObjectNotFound:
             raise StreamNotAvailable("Stream not available for this track")
-        except TooManyRequests:
-            raise TooManyRequests("Stream unavailable")
+        except TooManyRequests as e:
+            e.args = ("Stream unavailable",)
+            raise e
         else:
             json_obj = request.json()
             stream = self.requests.map_json(json_obj, parse=Stream().parse)
@@ -863,10 +870,12 @@ class Video(Media):
 
         try:
             request = self.requests.request("GET", "videos/%s" % self.id)
-        except ObjectNotFound:
-            raise ObjectNotFound("Video not found or unavailable")
-        except TooManyRequests:
-            raise TooManyRequests("Video unavailable")
+        except ObjectNotFound as e:
+            e.args = ("Video with id %s not found" % media_id,)
+            raise e
+        except TooManyRequests as e:
+            e.args = ("Video unavailable",)
+            raise e
         else:
             json_obj = request.json()
             video = self.requests.map_json(json_obj, parse=self.parse_video)
@@ -891,8 +900,9 @@ class Video(Media):
             )
         except ObjectNotFound:
             raise URLNotAvailable("URL not available for this video")
-        except TooManyRequests:
-            raise TooManyRequests("URL unavailable)")
+        except TooManyRequests as e:
+            e.args = ("URL unavailable",)
+            raise e
         else:
             json_obj = request.json()
             return cast(str, json_obj["urls"][0])